+++ app/transitions.js
@@ -1,0 +1,8 @@
+export default function () {
+  this.transition(
+    this.fromRoute('account.asset.index.index'),
+    this.toRoute('account.asset.index.new'),
+    this.use('toUp'),
+    this.reverse('toDown')
+  );
+}