+++ CHANGELOG.md
@@ -1,3 +1,10 @@
-- Added account/asset/index/new
+James - 8/9/16
+-
+- Add `transitions.js` file for liquid fire use
+- Add `isExpiring` cp to `auth` service and stub `refreshSession` for future use
+- You can now add attributes to asset categories
+- Add session expiration notice to application template. Does nothing but display for now.
+- Add `onResponse` action attr handler to `confirm-modal` component for alternate usage
+- Added `account/asset/index/new`
 - Added Array Transform
-- Added attributes key to asset model
+- Added `attributes` key to `asset` model